// CPL/admin-backend/libs/grpc-client/src/grid-trading/enum.ts
export enum STRATEGY_STATUS {
  PENDING = 1,
  RUNNING = 2,
  ENDING = 3,
  END = 4,
  ERROR = 5,
}

export enum GRID_TYPE {
  ARITHMETIC = 1,
  GEOMETRIC = 2,
}

export enum SORT_TYPE {
  DESC = 1,
  ASC = 2,
}

export enum SEARCH_BY_FIELD {
  BOT_ID = 1,
  EMAIL = 2,
}

export enum ACTIVE_COPY {
  ACTIVE = 1,
  INACTIVE = 2,
}